added support for PD values
git-svn-id: https://svn.grrrr.org/ext/py/trunk@2223 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/modmeth.cpp
@@ -227,7 +227,7 @@ PyObject *py::py_setvalue(PyObject *self,PyObject *args)
 	PyObject *val = PyTuple_GetItem(args,1); // borrowed reference
     if(name && val && PyString_Check(name) && PyNumber_Check(val)) {
 		const t_symbol *sym = MakeSymbol(PyString_AsString(name));
-        float f = PyFloat_AsDouble(val);
+        float f = (float)PyFloat_AsDouble(val);
 
         if(value_setfloat(const_cast<t_symbol *>(sym),f))
 		    post("py/pyext - Could not set value '%s'",GetString(sym));
